Fetch to Post new template from TemplatePage.tsx moved to separate file
