+++ frontend/src/pages/AddTemplate/TemplatePage.tsx
@@ -1,37 +1,15 @@
 import React, { useState } from 'react';
-import {Addresses} from '../../api/LocalHostAddresses';
+import { addCourseTemplate } from '../../api/addTemplateService';
 
 function AddTemplatePage() {
     const [inputText, setInputText] = useState('');
     const [displayedText, setDisplayedText] = useState('');
 
     // Function to handle the POST request
-    const sendPostRequest = async (courseTemplateName:string) => {
-        try {
-            const response = await fetch(`${Addresses.COURSETEMPLATES}/courseTemplates/add`, { // Replace with your endpoint
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-                body: JSON.stringify({
-                    name: courseTemplateName,
-                    id: '12345'
-                })
-            });
-            if (!response.ok) {
-                throw new Error(`HTTP error! status: ${response.status}`);
-            }
-            const data = await response.json();
-            console.log(data); // Process the response data as needed
-        } catch (error) {
-            console.error('There was an error!', error);
-        }
-    };
-
-    // Function to handle button click
     const handleButtonClick = () => {
         setDisplayedText(inputText);
-        sendPostRequest(inputText); // Send the POST request
+        addCourseTemplate(inputText) // Send the POST request
+            .catch(error => console.error('POST request failed: ', error));
     };
 
     return (
